Applied clippy lints

+++ node-graph/gstd/src/executor.rs
@@ -405,7 +405,7 @@ async fn blend_gpu_image(foreground: ImageFrame<Color>, background: ImageFrame<C
 		.create_compute_pass(
 			&pipeline,
 			Some(readback_buffer.clone()),
-			ComputePassDimensions::XY(background.image.width as u32, background.image.height as u32),
+			ComputePassDimensions::XY(background.image.width, background.image.height),
 		)
 		.unwrap();
 	executor.execute_compute_pipeline(compute_pass).unwrap();
